Add files via upload
Shop management system is designed to control all the activities of a Shop wherein we can add, update, delete, products for managing shop system. Also the top five sales are featured in the dashboard in order to have extra stock for that product . Then sales are managed. Also sales report for every date and interval of dates and monthly sales are tracked.